refactor: 메인페이지 파일 분리 (default/mobile) 및 공통 함수 분리
