+++ src/utils/user.ts
@@ -1,0 +1,14 @@
+export const handleLogin = () => {
+  /* TODO: LOGIN */
+  // setUser({ name: 'PIMFY' });
+};
+
+export const handleLogout = () => {
+  /* TODO: LOGOUT */
+  // setUser(undefined);
+};
+
+export const handleJoin = () => {
+  /* TODO: JOIN */
+  // setUser({ name: 'PIMFY' });
+};
